refactor as code review comments
