+++ README.md
@@ -13,8 +13,13 @@ python: true
 require: 
   - ./readme.python.md
   - ./readme.cli.md
-try-require:
-  - $(this-folder)/readme.az.common.md
+
+az_common:
+    ## The check-name-availability and list-operation are suppressed in CLI by default,
+    # enable_methods:
+    #   - check-name-availability
+    #   - list-operation
+    enable_methods: []
 
 pipeline-model: v3
