--- readme.az.common.md
@@ -1,14 +1,0 @@
-# configuration
-
-These settings apply only when `--az` is specified on the command line.
-
-``` yaml $(az)
-
-az_common:
-  ## The check-name-availability and list-operation are suppressed in CLI by default,
-  # enable_methods:
-  #   - check-name-availability
-  #   - list-operation
-  enable_methods: []
-
-```
