+++ src/plugins/azgenerator/TemplateAzureCliTestScenario.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { CodeModelAz } from "./CodeModelAz"
+import { PreparerEntity, get_resource_key } from "./scenario_tool"
 
 export function GenerateAzureCliTestScenario(model: CodeModelAz): string[] {
     let head: string[] = [];
@@ -85,8 +86,29 @@ export function GenerateAzureCliTestScenario(model: CodeModelAz): string[] {
 
     let imports: string[] = [];
     let decorators: string[] = [];
-    model.FormatPreparers(imports, decorators);
+    FormatPreparers(model, imports, decorators);
 
     return head.concat(imports, class_info, decorators, initiates, body);
 }
 
+function FormatPreparers(model: CodeModelAz, imports: string[], decorators: string[]) {
+    let decorated = [];
+    for (let entity of (model.GetPreparerEntities() as PreparerEntity[])) {
+        //ret.push("    @ResourceGroupPreparer(name_prefix='cli_test_" + m + "')")
+        let line: string = `    @${entity.info.name}(name_prefix='cli_test_${model.Extension_NameUnderscored}_${entity.object_name}', key='${get_resource_key(entity.info.class_name, entity.object_name)}'`;
+        for (let i = 0; i < entity.depend_parameter_values.length; i++) {
+            line += `, ${entity.info.depend_parameters[i]}='${entity.depend_parameter_values[i]}'`
+        }
+        line += ")";
+        decorators.push(line);
+        if (decorated.indexOf(entity.info.name) < 0) {
+            if (entity.info.name=='ResourceGroupPreparer') {
+                imports.push(`from azure.cli.testsdk import ${entity.info.name}`);
+            }
+            else {
+                imports.push(`from .preparers import ${entity.info.name}`);
+            }
+            decorated.push(entity.info.name);
+        }
+    }
+}
